Feature: Playbooks (#3)
Changes:

- ✨ feat: create a first contact playbook to scan and register hosts in the known_hosts file.
- ✨ feat: create a security playbook to create group, user and ssh key pair.
- ✨ feat: disable root login, password login and empty password login to force to use SSH key pairs.

+++ ansible-control-node-vm/sh/ansible.sh
@@ -3,6 +3,11 @@
 # Ansible
 echo "--------------------------------------------------------------------------------------"
 echo "Installing Ansible..."
+if [ -z "$1" ]; then
+  echo "You must supply Ansible vault secret!"
+  echo " ./ansible.sh 'SECRET'"
+  exit
+fi
 sudo apt remove ansible
 sudo apt --purge autoremove
 sudo apt update
@@ -10,7 +15,10 @@ sudo apt upgrade
 sudo apt -y install software-properties-common
 sudo apt install software-properties-common
 sudo apt install -y ansible
-sudo apt install -y sshpass
-echo "Editing Ansible Inventory..."
-sudo sh -c "echo '[servers]' > /etc/ansible/hosts"
-sudo sh -c "echo node-host ansible_host=192.168.59.3 ansible_ssh_user=vagrant ansible_ssh_common_args=\'-o StrictHostKeyChecking=no\' >> /etc/ansible/hosts"
+echo "Ansible Installed"
+echo "Creating Vault Password File..."
+sh -c "echo '$1' > $HOME/.vault_pass"
+echo "Vault Password File Created"
+echo "Setting Vault Password File in ansible.cfg..."
+sudo sed -i '/vault_password_file$/a vault_password_file = $HOME/.vault_pass' /etc/ansible/ansible.cfg
+echo "Vault Password File Set"
